feat: update tax report page
