+++ tinlake-ui/containers/PoolManagement/queries/index.ts
@@ -2,8 +2,9 @@ import { PoolData } from '../../../utils/usePool'
 import { assetList } from './AssetList'
 import { dailyInvestorBalances } from './DailyInvestorBalances'
 import { dailyTokenPrices } from './DailyTokenPrices'
-import { investorTransactions2020, investorTransactions2021 } from './InvestorTransactions'
+import { investorTransactions } from './InvestorTransactions'
 import { rawPoolData } from './PoolData'
+import { taxReport2020, taxReport2021 } from './TaxReports'
 
 export type PoolQuery = ({ poolId, poolData }: { poolId: string; poolData: PoolData }) => Promise<boolean>
 
@@ -12,8 +13,9 @@ const queries: { [name: string]: PoolQuery } = {
   'Daily token prices': dailyTokenPrices,
   'List of assets': assetList,
   'Raw pool data': rawPoolData,
-  'Tax report 2020': investorTransactions2020,
-  'Tax report 2021': investorTransactions2021,
+  'Investor transactions': investorTransactions,
+  'Tax report 2020': taxReport2020,
+  'Tax report 2021': taxReport2021,
 }
 
 export default queries
